backend/requirements.txt
# ./backend/requirements.txt
# Python dependencies for the SplitFair backend.

fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
pymysql
python-multipart
websockets

